⚡️ :: ChallengeParticipantsDTO 분할
